[transaction-controller] Rename `test` directory to `tests`
